Updated FIRSTS sets to support epsilon rules and fixed a few bugs
The calc_firsts algorithm now supports empty (epsilon) rules. This also
fixes a bug in the algorithm that adds the FIRSTs of all Nonterminals, instead of
just the first one. Only if a Nonterminal has an epislon rule, it should add
the FIRSTS of the succeeding Nonterminals to the current FIRST.

We also have to take care of Terminals in the second loop. So I'm wondering if
we can merge the two loops together now.