Fix subtle bug due to comparing a rule by value and not by position.
The basic problem is if one has a rule such as:

  A ::= B 'c' B

the FIRSTS calculation would look at the last B and say "is it the last
nonterminal in the alternative?" by comparing itself to the last nonterminal in
the alternative by value. In an example such as the above, the last nonterminal
just so happens to have the same value as a nonterminal before a terminal, so
this comparison is incorrect.

This patch makes a simple fix and introduces an appropriate test case.