Make the index types be usize.
The original thinking was that these types would be used a lot, and that
representing them as i32s would be a significant memory saving. In practise,
they're not likely to be used *that* much, and all that it ended up causing was
a lot of annoying casting.

This commit changes the types to usize and removes all the associated casting.
For convenience, it also explicitly stores nonterms_len and terms_len, which
make code which rely on that information a little clearer.