Move the stategraph builder to use ints and bitvecs.
Unlike the previous version this does not use a todo list. I believe that the
previous version could add the same state to the todo list multiple times,
making stategraph building exponential. By adding known new states straight to
the overall list of states, this operation should be roughly linear to the
number of terminals and nonterminals in the grammar.