Removed the yacc feature for redeclaring nonterminals as tokens.
In Yacc, one can redeclare a nonterminal in a rule to a token, by
adding the nonterminal name to the %token list. We won't support this
for now, as we are using that %token list to store all possible terminals
in the grammar. Terminals that are found during parsing of the grammar
are added automatically to that list.