Change the representation of items away from a single bitvec.
Instead of representing all the lookaheads of a set of items as a single
contiguous item, this breaks them up into one bitvec per item. This makes
reasoning about the data-structure a little easier, and also allows us to make
use of BitVec::union in some performance critical corners. It also requires less
memory to be allocated, as dots that aren't active require a single None (rather
than grm.terms_len bits).

On the Python grammar on my machine, this patch uses about 40% less memory and
is a bit over 2x faster.