Simplify the representation used by moving to sets of items.
In essence, this commit implements a more traditional, mildly optimised
Knuth-ish LR state generator. Itemsets are now HashSets with keys (alternative
offset, dot) mapping to contexts (i.e. lookaheads). This allows us to simplify
the implementation somewhat which is particularly noticeable in the goto()
function. The close() function has also grown a small optimisation (prompted by
Chen's dissertation), whereby we only check states that can plausibly have
changed (i.e. in general we don't have to continually rescan the whole itemsem).

However, more significantly we are able to get rid of the significant number of
RefCells of the previous version. These appeared necessary because we are often
iterating over a data-structure and modifying it at the same time. However,
because of the quantity of RefCells we were using, we were spending huge amounts
of time in reading/writing the borrowing state. This commit makes use of todo
lists, splitting up iterating over a data-structure (and adding things to a todo
list) from modifying it (putting things from the todo list into the
data-structure). On the Python grammar this leads to a roughly 8.5x speedup.