Implement much of Pager's weakly compatible optimisation.
As well as the basic weakly compatible optimisation, this commit also maintains
a list of plausible candidate matches. It is thus about 15x faster on the Python
grammar.

As well as the obvious weakly compatible changes, this commit also doesn't store
itemsets' closures, only calculating them as needed. This makes the weakly
compatible checks faster and easier to understand.

[The Pager test was created by Lukas Diekmann, and I have shamelessly cut and
paste that code from another branch.]