Create state table from a state graph.
For simplicities sake, we currently use two hashmaps to represent the actions
and gotos. It seems that in practise, both are quite densely filled (on a couple
of real grammars I tried, about 1/2 of the actions slots are filled), so it
might be worth making these multi-dimensional vectors. But that's an
optimisation to worry about later, if at all.