Add a simple GC for when unreachable states are created.
Surprisingly, the Pager paper doesn't talk about this, perhaps because it used
deterministic list iteration, whereas we often use nondeterministic set
iteration. Regardless, even in the example from Pager's paper, one can
occasionally create unreachable states (on perhaps 1 out of 100 runs, 24 or 25
states will be created instead of 23). We thus need to weed out unreachable
states and update edges accordingly.

Testing this is difficult, because on the non-determinism. I have been unable to
come up with a test that deterministically triggers state GC. I have therefore
resorted to running the Pager test a large number of times, on the basis that if
there is a bug in the GC, it will trigger during one of those runs. This seems a
bit evil, but I can't think of a better way of doing things.