Move close from a set todo to a vector.
Doing the equivalent of pop() on a set is so slow, and the number of items so
small (for example, on the Java grammar, the median number of items in the todo
is 8), that we lose all the normal performance advantages of sets in this case.

On the Java grammar this speeds us up by about 10% overall.