Only calculate closed states when needed.
Since the same state can be added to the todo list multiple times before it is
actually processed, we were previously creating closed sets that were never
actually used. This commit ensures that closed states are only created when
they're going to be used. On the PHP grammar, this speeds things up by about
10%.