Deal with Rust Strings bytes / codepoints conflation.
As soon as one uses multi-byte unicode symbols (e.g. ❤, which takes two bytes),
then one notices that these &""[x..y] is doing something different than
&"".chars().take(x). The former works on bytes, the latter on codepoints (what
we would typically call characters). I had naively assumed that both operate on
codepoints.

This patch fixes this by being clear that the indexes it is passing around are
always bytes. So if one wants to access a char at byte 'i' one has to do
&s[i..].chars().next(). This feels a bit clumsy, but is much faster than the
(incorrect!) &s.chars().skip(i).next() which has to do a linear scan over the
whole string to find the ith character; in other words, the cost of obtaining
this is O(n), not O(1) as expected.

Because of that, this patch has the pleasant side-effect of speeding up building
the PHP grammar by about 15-18%!