Move from rule IDs to token IDs.
Previously we tried using the index of a rule in a (frankly, somewhat arbitrary)
vector. This only makes sense when the lexer is entirely in control of the
possible rules. This is not true for many parsers which will add a separate end
token type (and, possibly, a start token type). It therefore makes a lot more
sense to always think of things in terms of token IDs.