Find the longest, and not the first, match.
Previously if we had these two rules:

  if IF
  [a-z]+ ID

then the input "iff" would produce (IF, if) (ID, f). The lex spec (which I should have
read more carefully!) sensibly stipulates that one should always prefer the longest
match (falling back on first-rule-in-the-file-wins only as a differentiator
between matches of the same length). This patch fixes this, so that "iff"
produces (ID, IFF).