Make PIdx its own distinct type.
Currently all the *Idx types are simple aliases for usize; that means that from
rustc's perspective an RIdx is the same as an SIdx is the same as a PIdx. This
is an easy way to get caught out. It also means that we tend to leak lots of
details about our internal representation to the outside world (here's a usize!
do what you want with it!).

This commit makes PIdx a distinct type which has to be explicitly converted
to/from a usize. It's no longer possible to confuse PIdx with another *Idx type
without doing something explicit. Interestingly, this already pointed out one --
admittedly completely benign -- case where the original types were used
incorrectly: in one place I used PIdx where I should have use RIdx. This commit
allowed the compiler to catch the error.