Allow users to specify the types of token IDs.
There's no reason to force this to be a usize (which, generally, is massive
overkill). Abstracting this detail away opens up the potential for a fun
optimisation.

Unfortunately, this commit really needs TryFrom, which is only in Rust nightly;
so, until it stabilises (https://github.com/rust-lang/rust/issues/33417), we're
dependent on Rust nightly (not stable). Oh well.