Internally make the size of a lexeme a u32.
This allows us to pack a Lexeme (assuming its TokId is a u32 or smaller) into
two machine words rather than three as before. It does mean that each lexeme is
limited to 4GiB in length, although in practise that's probably quite large
enough.

Going forward, we could be much more clever about doing low-level packing, but
this is probably good enough for the time being.