Before state merging, try finding an exactly equivalent state.
It's not clear that one needs this with Pager's algorithm, but without it, one
can't turn the weakly compatible check off (since the program endlessly creates
more and more states). When we do find an exactly equivalent state, we also have
the advantage that we don't need to do any change propagation, so it's probably
a small optimisation as well. [FWIW, this is clearly what Hyacc does; Chen makes
a vague mention to "the same" states in his thesis, which is where I got the
idea from.]