This docstring was broken.
First, it wasn't actually executing as a doc test. Second, it had an incorrect
assertion:

  assert!(firsts.is_epsilon_set(grm.nonterminal_off("S")));

However, since Firsts is private to stategraph.rs, testing it as a doc test is
impossible (there's no way to "use" Firsts). So there's no point pretending this
comment is anything other than a bog-standard comment. So turn the doc test into
a comment and fix the assertion.