Separate lexer definitions from lexers.
This is a step on the road to something bigger.

Previously the Lexer struct spat out a Vec of Lexemes, meaning that there was no
way to ask questions of Lexemes such as "what line number did you start with"
(since Lexemes are, intentionally, small at two machines words on a 64 bit
machine). The only way to do this is to have an intermediate result struct which
contains this sort of global-ish information.

This commit does precisely that: Lexer becomes LexerDef; on then makes a Lexer
(specific to a given input) from LexerDef; and ask the Lexer for a Vec of
Lexemes.