When setting rule IDs, calculate which tokens are missing from lexer / parser.
Tokens defined in the lexer but not referenced in the parser are quite probably
harmless (or, at most, worth warning the user about). But tokens referenced in
the parser but not defined in the lexer are highly indicative of an error.
Calculating these and returning them to the caller allows them to make a choice
as to whether they want to do something with this information or not.