Remove the explicit representation of the EOF term.
It's now clear that explicitly representation of this is problematic: the user
doesn't specify it in their grammar, and even our rewritten grammar doesn't
explicitly reference it. Thus we had built in an ugly "get all terminals (note:
this excludes the EOF terminal" hack, which was an eyesore (and confusing). This
commit removes the explicit EOF terminal entirely.