Deal with the lack of an explicit EOF terminal.
cfgrammar no longer explicitly represents the EOF terminal: that has knock-on
effects which this commit deals with. The coping mechanism is very simple:
whenever we deal with a setting for all terminals, we add 1 for the EOF term
(hence the proliferation of "grm.terms_len() + 1" you'll see in this commit).