(Re)add an explicit EOF terminal to Yacc grammars.
Previously every terminal had a name and index: this made the EOF terminal very
confusing (it had a name the user hadn't specified, and thus showed up in
various places the user wouldn't expect). I then made a PR which removed the EOF
terminal entirely.

That then turns out to make error reporting annoying. Ideally, one wants to
record the terminal where an error happened, but if it happens at the end of
the file, what terminal should it be associated with?

The challenge then is to find a way of having an EOF terminal as part of the
grammar, but hide it in a natural way from end-user activities. This commit does
just that. The trick is to say that while every user-specified terminal must
have a name, terminals added by cfgrammar don't. In other words, this means that
if we say "give me every terminal with a name" we naturally get everything but
the EOF terminal.