Make StateGraph states/edges functions.
I should probably have done this a long time ago, but when StateGraph was purely
internal to lrtable that didn't seem important. It's now pretty clear that some
users will want access to the StateGraph as well as the StateTable, so making
this more amenable to future evolution is important.