Store, and allow access to, core as well as closed states.
This means, mostly, renaming functions which access closed states to be explicit
that they're access closed states; and then adding simple counterparts that
access core states.

Internally, storing this is pretty easy. Instead of storing a vector of closed
states, we store a vector of (core state, closed state) pairs.