Don't sort non-terminal names in the grammar.
This serves no real purpose, and is slightly confusing for users, who may
reasonably expect that the order of non-terminals isn't changed by cfgrammar
unless explicitly asked to. This commit sort-of maintains the original order of
the user's input. So if the grammar is:

  Z: ...;
  Y: ...;
  X: ...;

then asking for the sequence of rules will return Z, Y, X in order. However, if
the grammar is:

  Z: ...;
  Y: ...;
  X: ...;
  Z: ...;

i.e. additional productions are subsequently added to Z, then the order returned
will still be Z, Y, X. If the user wants to perfectly recreate such a grammar,
they'd have to peek inside the indexes of individual productions (which is
discouraged for other reasons). I think that's a sufficiently niche interest
that it's not worth thinking about, at least not at the moment.