Implement Corcheulo et al.'s recovery mechanism.
This is taken from "Repairing syntax errors in LR parsers" by Rafael Corchuelo,
Jose A. Perez, Antonio Ruiz, and Miguel Toro ("Corcheulo et al." henceforth).
This is a somewhat straight-ish implementation of the paper which, when a
parsing error is encountered, tries to generate repairs which allow the parser
to continue.

Given the Java grammar and this input file:

  class C { int

then running things gives the following output:

  $ lrpar java.l java.y tiny.java
  Error detected at line 1 col 14. Amongst the valid repairs are:
    Insert "IDENTIFIER", Insert "SEMICOLON", Insert "RBRACE"

In many cases the recovery mechanism is highly effective. However it is
non-deterministic. For example given:

  class C { int x = x.

Different runs find repairs in different orders:

  $ lrpar java.l java.y tiny.java
  Error detected at line 1 col 21. Amongst the valid repairs are:
    Insert "CLASS", Insert "SEMICOLON", Insert "RBRACE"
    Insert "IDENTIFIER", Insert "SEMICOLON", Insert "RBRACE"
    Insert "THIS", Insert "SEMICOLON", Insert "RBRACE"
  $ lrpar java.l java.y tiny.java
  Error detected at line 1 col 21. Amongst the valid repairs are:
    Insert "CLASS", Insert "SEMICOLON", Insert "RBRACE"
    Insert "IDENTIFIER", Insert "SEMICOLON", Insert "RBRACE"
    Insert "THIS", Insert "SEMICOLON", Insert "RBRACE"

This is fun because one of the repairs is arbitrarily used to continue parsing:
so on some runs a better quality repair is chosen than on others (the recovery
mechanism has only a very local view of the repair's effects, so is often
unaware of knock-on effects much later in a file).

This also combines with another factor: the recovery mechanism can sometimes
explode. This seems to happen with large grammars with large numbers of
terminals.

So, this is imperfect, but it's hopefully a gateway to something better.