Move the recovery mechanism over to using a cactus stack.
The previous mechanism ended up copying the parse tree when trying recovery
options (because each sub-parse and todo entry required a copy of the parse
tree). If an error was encountered near the end of a large file, this could end
up copying vast quantities of memory, which could (and did) lead to gigabytes of
RAM being consumed before memory eventually ran out. By using a cactus stack, we
avoid this problem. That's not to say that this is a memory-friendly approach:
it can still gobble up hundreds of megabytes, but I can no longer make my
standard example crash.