Aggressively flush the todo list when a new low score is found.
Often when we find a new successful sequence of repairs, it means that many
entries in the todo list can never succeed as their score is already too high.
By removing these entries as soon as we notice this, we get a useful speed up
(~10%) and memory saving (~25%) on nasty examples. There is no measurable effect
on smaller examples.