Ensure that the parse-tree parser doesn't parse too much.
There is potential here for an inefficient case to happen: if a parse ends with
a shift, then the parser could try parsing for PARSE_AT_LEAST symbols, even
though it should really only parse 1 symbol. This isn't a correctness issue, but
since it's undesirable, it might as well be fixed.