Return a parse tree even when an error has been encountered.
The resulting parse tree is a little odd, because some lexemes have types but no
value. So given a simple calculator-esque grammar and the input "(nn" one ends
up with trees like this:

 E
  E
   OPEN_BRACKET (
   E
    N n
   CLOSE_BRACKET
  PLUS
  N n

At first glance that might seem normal, but if the user had input "(n)+n" the
parse tree would look like:

 E
  E
   OPEN_BRACKET (
   E
    N n
   CLOSE_BRACKET )
  PLUS +
  N n

Still, it's up to the user whether they want to use the resulting parse tree. If
the only lexemes affected are things where one doesn't care about the value
(e.g. CLOSE_BRACKET or PLUS), it may be possible to do useful things with the
resulting tree.