Make the actions more obviously a sparse table (using a hashmap).
This is a simple representational shift: we move from having keys of the form
(StIdx, Symbol) to usize. Thus we give each cell in the statetable a unique
number (which is trivial, because we know that each row has precisely
grm.terms_len() elements).

This has no observable effect on normal parsing, but for clients of the
state_actions function it makes a huge difference (making it several orders of
magnitude faster).

Note: there are clearly better data-structures than we're using, but this is a
quick, easy win.