Correctly handle multiple implicit tokens.
Assume you specified:

  %implicit_tokens a b

This would lead to a rule of:

  ~ : "a" | "b" | ;

Which meant that at any point you could specify an "a" *or* a "b" -- but not a
sequence of "a"s and "b"s. This commit rewrites the rule so that it looks as
follows:

  ~ : "a" ~ | "b" ~ | ;

This allows arbitrary sequences of (possibly interleaved) "a"s and "b"s.