Rename validate to complete_and_validate, and juggle code accordingly.
This has always been an icky part of cfgrammar's internals: you had to call
validate after parsing, but the API slightly melted together the parser and
validator, and it wasn't really clear how they related. Since in the future it
seems highly likely that this function will need to perform mutation, make it
take "&mut self" as the first argument (even though we don't currently mutate
anything).

After moving this around, we can then make the "ast" function consume the
parser, making it clear that after parsing you get an AST and the parser is then
redundant.

These changes have various fairly mechanical impacts simply to get things
compiling together. Although it might not look it, this commit is about as
minimal as I can work out how to make it.