Add a minimum sentence generator.
For any given non-terminal this returns a (non-deterministic) minimal matching
string (note that, in general, there can be more than one matching string).

Since this is a somewhat specialist use case, we make this a lazy part of a
grammar: only when you want to generate minimal sentences do you construct a
sentence generator.