Implement the KimYi recovery algorithm.
After playing with this for weeks, I think I've understood the paper well enough
to derive a fairly faithful implementation (bearing in mind that the paper is
somewhat incomplete, and frequently confusing). As the paper suggests, this only
finds a single minimal cost solution to each parser error, so it's highly
non-deterministic: sometimes the minimal cost solution it comes up with is good,
and sometimes it isn't. It is, however, pretty fast, certainly compared to the
Corchuelo et al. algorithm. The biggest extension relative to the paper is that
they really only care about finding a repair: they don't seem to care about
reporting that to the user. The repairs thus reference non-terminals in the
grammar which, to most users, are completely incomprehensible. This commit uses
cfgrammar's SentenceGenerator to turn a non-terminal repair into a (possible
set) of terminal inserts, which is infinitely easier to understand. Thus whilst
KimYi might say a repair is:

  InsertNonterm expr

this commit will say something like:

  InsertTerm {Var, Identifier}, InsertTerm{+, -, *}, InsertTerm {Var, Identifier}

There are several glaring inefficiencies in this commit: notably, we continually
make new SentenceGenerators. That's because, currently, we don't really have a
good way for passing state between, and within, error recovery instances.
However, that's a more invasive change which can come later.