Better reflect the paper's algorithm.
I couldn't work out why, occasionally, repairs wouldn't be found. Then I found a
little sentence on p11:

  However, our guessed distance is 0 if t = 1 because reduce actions or
  deletions can possibly intervene for (q1...qi, p, t, e, c) to reach the goals.

What this means is that the R3I-S reduction rule at the top of p11 is
misleading: it should really be 2 separate rules, one for t==0 and one for t>1.
This commit updates things to reflect the prose.