First version of our new KimYi+ algorithm.
Whereas KimYi just tries to find a single repair, this variant finds all valid
minimal repairs.

As things stand, this isn't unusably slow (perhaps a couple of seconds for
complex cases), but it's not hugely fast either. The code could definitely do
with some optimising before it's ready for the big bad world; but that's best
done later.