Update the API to deal with the case where we can't find any repairs.
Some errors are so bad that we can't find suitable repairs, at which point it's
impossible to return a parse tree (it would, at best, be only a partial tree).
It therefore seems safest to say that, when an error occurs, the user may, or
may not, get back a parse tree too.