Throw an error if lexemes that aren't referenced in the parser are encountered.
As such, lrlex's original behaviour was defensible: it's not really its problem
whether a lexing rule is used by the parser or not. *But* it causes a subtle
problem with the ID mapping: if the parser doesn't say that rule R has id X,
then rule R ends up with a completely random id that baffles the parser. While
we could come up with a complex API to allow the parser to deal with such
lexemes, they're rare enough that the complexity seems worthless: better to have
the lexer explicitly choke on such lexemes as soon as they're encountered.