Allow the different recovery algorithms to have state.
Some of the recoverers (a word I think I've just coined) can use pre-calculated
data to speed up their actions. At the moment, there's no sensible way for them
to store this state, so they often calculate it over and over again.

This commit minimally moves the current 3 recoverers to use trait objects (via
the Recoverer trait). None of them currently make much use of that facility yet:
that will be for a later commit. Also note that, to help make the diff as easy
to review as is practical, I have not indented the code that now finds itself in
a trait. That will be for the next commit.