Completely rethink the MF recoverer.
KimYiPlus (and, indeed, KimYi) is flawed. Given a grammar such as:

  S: T U 'c';
  T: 'a';
  U: 'b';

and the input "c" it fails to find any repairs. The problem, I believe, all
stems from the mysterious "t" in the reduction rules on p11/12. The text says
this:

  Hence, our guessed distance using dist(q, a) is as follows. Our guessed
  distance is dist(qi, ap) if t > 1 for configuration (q1...qi, p, t, e, c)
  because dist(qi, ap) is the real distance to the least-cost goal in this case.
  However, our guessed distance is 0 if t = 1 because reduce actions or
  deletions can possibly intervene for (q1...qi, p, t, e, c) to reach the goals.

First of all, this means that the reduction rule R3I-S is wrong. It is written
as:

   q_i \xrightarrow[X] \wedge dist(q, a_p) < \infty
   -------------------------------------------------------------------
   (q_1 \ldots q_i, p, t, e, c_f, c_g)
   \xrightarrow[R3]
   (q_1 \ldots q_i \cdot q, p, t + 1, e \cdot (ins X), c_f + ic(X), d)

but I believe it should be:

   (t == 1 \wedge d = 0)
   \vee
   (t > 1 \wedge q_i \xrightarrow[X]
    \wedge  dist(q, a_p) < \infty
    \wedge d = dist(q, a_p))
   ------------------------------------------------------------------
   (q_1 \ldots q_i, p, t, e, c_f, c_g)
   \xrightarrow[R3]
   (q_1 \ldots q_i \cdot q, p, t + 1, e \cdot (ins X), c_f + ic(X), d)

This starts to make the root of the problem clearer: what is t really doing?
First of all, there doesn't appear any reason for it to be an integer: it's
basically a boolean in disguise. Second, there's an obvious (well, I say that
after 2 months of looking at this...) misinteraction with R3I-R, which has as a
condition:

   (q_1 \ldots q_i, p, 1, e, c_f, c_g)

In other words, as soon as the algorithm finds a possible chain of insertions,
it won't try doing any LR reductions. That explains why our example doesn't
work: the dist to "c" is 1, which requires inserting 2 tokens. However, those
two tokens are "hidden" by a reference to another rule; in order to be "found"
they require an LR reduction. But KimYi's algorithm doesn't allow that because
the insertion chain forces t to be greater than 1.

However, that almost looks like it doesn't matter, because R3I-R is designed to
insert non-terminals. However, that only works if the whole of the rest of the
production needs to be added. In other words if you're in a state with a
production:

  R: A . 'b' 'c'

then R3I-R will insert both 'b' and 'c'. If your input contains neither, fine;
but if it contains 'c', then R3I-R "overshoots" (it's almost trying to duplicate
'c').

If I had to guess, and was a bit cynical, I think that the problematic 't' was
necessary to get the single example in the KimYi paper working. But I might be
wrong. Whatever the reason for it, it's clearly a problem.

This commit tries an entirely new approach. It's currently a bit messy; it's
slow; and occasionally it duplicates answers; but it seems to work.

The first thing it does is to remove 't' entirely and with it any checks on 't'.

The second thing it does is to change the 'dist' function. The KimYi version of
this only considers forward edges in the stategraph which means that it doesn't
take into account reductions and gotos that could happen. This commit adds a new
dist function which (admittedly rather slowly) takes into account the possibly
recursive chain of reductions and gotos that could happen at a given state.
MF::dist always returns a distance less than or equal to the distance given
by KimYi::dist.

Going back to the example at the top, the MF recoverer correctly reports the
following repair:

  Insert "a", Insert "b"

A final, interesting, example shows that KimYi sometimes incorrectly reports
long repairs. Given this grammar:

  S:  'a' 'b' 'd' | 'a' 'b' 'c' 'a' 'a' 'd';

and the input "acd", KimYi incorrectly reports the following repair:

  Insert "b", Shift, Insert "a", Insert "a"

whereas MF correctly reports the following:

  Insert "b", Delete

Is MF perfect? Dunno. For example, I have a sneaking feeling that it should
dynamically take into account tokens from p + 2 in the input to handle Deletes,
but I have been unable to craft an input which shows this, so perhaps I'm wrong.