Allow empty strings to be repaired.
This requires change #3f6d7061 from lrtable, which exposes the statetable's
final state. From that, we can add the distance to the EOF terminal to the dist
table.

Doing this also nicely exposes one of the current inefficiencies in MF in an
existing test: duplicate repairs in the face on non-terminals which are
completely repaired.