Lazily create the min/max costs.
If someone only wants a minimum length sentence, they don't want to pay the
(potentially noticeable) costs of working out the necessary data for maximum
length sentences.

The downside to this is that RefCell's borrowing is not cheap. It may be
necessary to switch to something like an UnsafeCell to reduce those costs if
they turn out to start hurting.