Add our own astar function.
The pathfinding crate has never been an entirely perfect fit for our purposes:
it's too general and does things we don't need to do, which meant that I always
suspected it was a bit slower than we needed. That was exacerbated when I tried
to updated to pathfinding 0.6.8 and found out that it slowed us down by 5-10x
(as well as changed the API).

This commit adds a customised astar_all function. It is simplified in two
significant ways over pathfinding's astar_bag:

  1) astar_all doesn't store paths: it just returns success nodes. That's fine
     for us, because we can retrieve the path from the repair cactus in each
     node. Ridding ourselves of this responsibility means that we get rid of a
     lot of parent pointer chasing that astar_bag has to deal with.

  2) astar_all is less concerned about duplicates. It can be that we reach the
     same node via two different routes: we then only want to keep the lowest
     cost of these. astar_bag has to keep hold of all the nodes it's ever seen
     in order that it can guarantee it doesn't search the same path twice. This
     isn't a concern for us: because we use repairs as part of a node's
     identity, we can't reach the same node twice anyway. Even if we could (and
     future changes make this more likely), my quick experiments suggest that we
     don't very often reach duplicate nodes anyway, so the costs might not be
     worth the hassle. [I'm not going to stake my life on this yet, but it's not
     completely implausible.]

The good news is that this significantly speeds up my standard examples by at
least 2-5x (bigger examples winning more), getting our running time on those
examples down to something acceptable.