We no longer need to insert non-terminals.
Because our dist() function correctly (well, hopefully!) deals with reductions
and gotos, the KimYi r3ir rule doesn't make sense, as it causes duplicates in
our searches.

This version of r3ir only reduces states when the dot is at the end of the rule.
In an ideal world, it might seem that we'd want to adjust lr_cactus to do this
for us, but it's possible that a future optimisation might bring back some
variant of r3ir, so I'm going to keep it hanging around for a while.

This does lead to a speed-up, which is obviously wildly dependent on grammar
and input -- but 25% seems not uncommon.