Use smaller types for term and nonterm costs.
In retrospect I chose ludicrously large types. Why would I want to give a
terminal a cost of u64::max_value() or thereabouts? I have no idea. To be
honest, I suspect that in general the scoring of terminals is likely to be in
the range of 1--10, and probably towards the lower end of that range.

On that basis, this commit changes things to only allow terminals to be scored
with a u8. This then makes it possible to reduce the score type of a nonterminal
to a u32. A rule would need to contain over 16 million terminals with a maximum
score to get close to overflowing this which seems, to put it mildly, rather
unlikely.

All in all, this commit helps reduce memory usage significantly; perhaps not so
much in cfgrammar itself, but in other code which uses the costs produced here.
It also gives people a clearer indication about the likely values they're likely
to encounter in real life.