Minimally move MF's sub-functions into the main impl.
This commit is as mechanical as possible: it just moves the code and does the
minimum to have it compile. As a happy bonus, this also simplifies the use of
traits (now one trait definition per impl, rather than one per function
previously).

In the near future, it will allow functions to share references to things like
the parser without having to have everything passed as an explicit argument.