Use a dynamic, rather than static, notion of distance.
In KimYi's approach (which, even though we've extended, we've maintained up
until now), the distance from a state to a lexeme is a static property, dictated
ultimately by the user's grammar. This meant that we could only safely apply
distances to sequences of insertions, because deletions could otherwise
invalidate the static calculations.

This commit, though small, fundamentally changes this. As well as the static
distances, we also consider dynamic properties of the user's input and the
repairs we've found thus far.

The most obvious example is the following. Imagine that we want to ask the
distance from state st_idx to lexeme T. Let's assume the static distance is 2.
Imagine T is immediately followed by another lexeme U which has a static
distance from st_idx of 0. Let's also assume both T and U have a term_cost of 1.
If we delete T then the distance to U is 1 (the deletion cost of T + the static
distance to U), which is a shorter distance than T. We therefore need to return
a distance of 1, even though that is the distance to the second lexeme. [There's
another case involving shifts that's explained in a code comments.]

This has a nice effect: we can now use (dynamic) distances for all nodes,
because we never have to worry that we're going to accidentally overestimate the
distance. This is a very helpful performance boost (about 30% on a few of my
standard examples).