Simplify next_lexeme to only return a Lexeme.
There is only one caller that needs to make use of the lexeme_prefix
functionality, so it's best to make the caller take that burden and remove that
parameter. Since next_tidx encompasses the second part of the return tuple, that
means that we can then simplify the return type to simply be a Lexeme. This
makes code that calls this function that little bit easier to read.