Move the astar algorithm from using sets to lists.
Because we no longer have any duplicate repairs, there is no need to use sets
for storing the astar todo list: a simple list has exactly the same effect. This
then saves us the overheads of hashing and hashmap storage that comes with sets.
On my torturous example it leads to a 24% performance increase.