Don't add (Insert, Shift) and (Shift, Insert) repairs when possible.
If we're about to insert term T and the next terminal in the user's input is T,
we could potentially end up with two similar repair sequences:

  Insert T, Shift
  Shift, Insert T

From a user's perspective, both of those are equivalent. From our point of view,
the duplication is inefficient: we end up carrying around two almost-identical
repair sequences which doubles the search space.

From our perspective, the two sequences aren't equivalent: we actually prefer
the first repair sequence. This is because one way of finding success nodes is
to have the repair sequence end in PARSE_AT_LEAST consecutive shifts. The first
repair sequence makes it more likely that we'll find those consecutive shifts.

This commit therefore adds a simple check that says "if I'm about to insert T,
check that the previous repair wasn't a shift of T; if it was, don't do the
insert". For many inputs this makes no difference; on one of my nasty examples,
it speeds things up by 5%. More importantly it means that instead of getting
this sort of output from lrpar:

  Found two repairs:
    Insert T, Shift
    Shift, Insert T

we now only get:

  Found one repairs:
    Insert T, Shift

which makes life easier for the user.